Lines now move when point does ANTLR updated to 4.2-SNAPSHOT to work with Java 8 Multifunctions update to work with JavaFX

+++ src/main/java/shiro/interpreter/ShiroBasePassListener.java
@@ -98,7 +98,7 @@ public class ShiroBasePassListener extends ShiroBaseListener {
     @Override
     public void exitNumberExp(ShiroParser.NumberExpContext ctx) {
         shiro.expressions.Number n = new shiro.expressions.Number(
-                Float.parseFloat(ctx.NUMBER().getText()));
+                Double.parseDouble(ctx.NUMBER().getText()));
 
         // save the result for later
         setExpr(ctx, n);
